requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1

# Computer Vision
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74

# Machine Learning
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.0.0

# ONVIF (optional)
onvif-zeep==0.2.12

# Database
SQLAlchemy==2.0.19

# Utilities
python-dotenv==1.0.0
